--- lib/pbdrv/drv/rp2040/mod.h
@@ -1,13 +1,0 @@
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-//! puzzle bus driver setup
-void pbdrv_setup();
-
-#ifdef __cplusplus
-}
-#endif
-
